// packages/codemarie/src/shared/proto/host/window.ts
export enum ShowMessageType {
	ERROR = 0,
	INFORMATION = 1,
	WARNING = 2,
	UNRECOGNIZED = -1,
}